% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwise_search.R
\name{backward_check}
\alias{backward_check}
\title{Backward search}
\usage{
backward_check(ESet, glob.f, glob.f.ks, m, alt, wts)
}
\arguments{
\item{ESet}{an ordered expression set object with the same sample ordering and features as processed by the ks.stepwise() function when performing a step-wise heuristic search}

\item{glob.f}{a vector containing the feature names whose union gives the best score (so far) in the search. Feature names should match those of the provided expression set object}

\item{glob.f.ks}{KS score (p-value or statistic depending on chosen metric of interest) corresponding to the union of the specified vector of features}

\item{m}{a character string specifying which metric to use for stepwise search criteria. One of either 'pval' or 'stat' may be used, corresponding to the KS p-value or statistic. Uses value passed in the ks.stepwise() function}

\item{alt}{a character string specifying the alternative hypothesis, must be one of "two.sided","greater" or "less". Uses value passed in the ks.stepwise() function}

\item{wts}{a vector of weights to use if performing a weighted-KS test. Default is NULL, passed in the ks.stepwise() function}
}
\value{
A list object with the first element being the set of features that gave a better score than the existing best score, and second being the corresponding score (to update the KS search results)
}
\description{
Performs a 'leave-one-out' evaluation on the existing best feature-set to see if the removal of any pre-selected feature gives a better score than that of the existing metafeature obtained in the step-wise search
}

