% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRCA_GISTIC_MUT_SIG.R
\docType{data}
\name{BRCA_GISTIC_MUT_SIG}
\alias{BRCA_GISTIC_MUT_SIG}
\title{Genomic Data from TCGA BRCA that associated with YAP/TAZ Activity}
\format{
An object of class \code{ExpressionSet} from \code{Biobase} package
containing a matrix of 16,873 rows (features) and 963 columns (samples) 
see (\href{http://bit.ly/2rjy29l}{ExpressionSet documentation})
}
\usage{
data(BRCA_GISTIC_MUT_SIG)
}
\description{
An ExpressionSet object consists of 16,873 genomic features across 963 samples.
}
\references{
Kartha VK, Kern JG, Sebastiani P, Zhang L,
Varelas X, Monti S (2017) CaDrA: A computational framework for performing 
candidate driver analyses using binary genomic features. 
(\href{https://www.frontiersin.org/articles/10.3389/fgene.2019.00121/full}{Frontiers in Genetics})
}
\keyword{datasets}
