\name{CaDrA-package}
\alias{CaDrA-package}
\docType{package}
\title{
Candidate Driver Analysis
}
\description{
Supports the identification of candidate (epi)genetic drivers using a step-wise heuristic search approach given ranked genomic data.
}
\details{
\tabular{ll}{
Package: \tab CaDrA\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-11-24\cr
License: \tab MIT\cr
}
library(CaDrA)
}
\author{
Vinay Kartha
Stefano Monti

Maintainer: Vinay Kartha <vkartha@bu.edu>
}
\references{
Kartha VK, Kern JG, Sebastiani P, Zhang L,
Varelas X, Monti S (2017) CaDrA: A computational framework for performing
candidate driver analyses using binary genomic features.
(\href{https://www.biorxiv.org/content/early/2017/11/23/221846}{bioRxiv})
}
\keyword{ package }