% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{CaDrA_Server}
\alias{CaDrA_Server}
\title{Shiny Server modules}
\usage{
CaDrA_Server(id)
}
\arguments{
\item{id}{A unique namespace identifier}
}
\value{
Shiny server elements
}
\description{
Shiny Server modules
}
\examples{

# Load R library
library(shiny)

# Create ui and server for Shiny app
id <- "myapp"

ui <- fluidPage(
   CaDrA::CaDrA_UI(id = id)
)

server <- function(input, output, session) {
   CaDrA::CaDrA_Server(id = id)
}

app <-  shinyApp(ui=ui, server=server)

# Launch and deploy Shiny app
# shiny::runApp(app, host='0.0.0.0', port=3838)
 
}
