% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{CaDrA_UI}
\alias{CaDrA_UI}
\title{Shiny UI modules}
\usage{
CaDrA_UI(id)
}
\arguments{
\item{id}{A unique namespace identifier}
}
\value{
Shiny ui elements
}
\description{
Shiny UI modules
}
\examples{

# Load R library
library(shiny)

# Create ui and server for Shiny app
id <- "myapp"

ui <- fluidPage(
   CaDrA::CaDrA_UI(id = id)
)

server <- function(input, output, session) {
   CaDrA::CaDrA_Server(id = id)
}

app <-  shinyApp(ui=ui, server=server)

# Launch Shiny app (NOT RUN)
# shiny::runApp(app, host='0.0.0.0', port=3838)

}
