% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCLE_MUT_SCNA.R
\docType{data}
\name{CCLE_MUT_SCNA}
\alias{CCLE_MUT_SCNA}
\title{Binary Genomic Data}
\format{
An object of class \code{ExpressionSet} from \code{Biobase} package
containing a matrix of 17723 rows (features) and 82 columns (samples) 
see (\href{http://bit.ly/2rjy29l}{ExpressionSet documentation})
}
\usage{
data(CCLE_MUT_SCNA)
}
\description{
An actual feature matrix comprises of 17723 features where 1/0 vectors 
indicating the presence/absence of the feature in the samples. There
are 82 samples in this study.
}
\references{
Kartha VK, Kern JG, Sebastiani P, Zhang L,
Varelas X, Monti S (2017) CaDrA: A computational framework for performing 
candidate driver analyses using binary genomic features. 
(\href{https://www.biorxiv.org/content/early/2017/11/23/221846}{bioRxiv})
}
\keyword{datasets}
