% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_score.R
\name{compute_score}
\alias{compute_score}
\title{Compute skewdness scores per feature}
\usage{
compute_score(mat, method = c("ks", "wilcox"), ...)
}
\arguments{
\item{mat}{matrix of binary features to compute row-wise ks scores for}

\item{method}{a character string specifying the method used to score features, must be one of "ks" or "wilcox"}

\item{...}{additional arguments provided to either ks.genescore() or wilcox.genescore() functions depending on method of choice}
}
\description{
Compute scores based on skewdness of a given binary matrix (based on sample ordering) used in the stepwise search
}
