% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knnmi.R
\name{cond_mutual_inf_ccc_vec}
\alias{cond_mutual_inf_ccc_vec}
\title{Conditional mutual information estimation for continuous case of 3 vectors.}
\usage{
cond_mutual_inf_ccc_vec(x, y, z, k = 3L, seed = 0L)
}
\arguments{
\item{x}{input vector.}

\item{y}{input vector of the same length as x.}

\item{z}{conditional input vector of the same length as x.}

\item{k}{number of nearest neighbors.}

\item{seed}{integer random seed. If seed<=0 a random seed is used.}
}
\value{
a double-precision value - mutual information estimation for
vectors \code{x} and \code{y}.
}
\description{
Compute conditional mutual information of \code{x},\code{y} given \code{z}
where \code{x}, \code{y} and \code{z} are all continuous
}
\examples{
data(mutual_info_df)

cond_mutual_inf_ccc_vec(mutual_info_df$Zc_XcYc,
                       mutual_info_df$Xc, mutual_info_df$Yc)
## 0.2936858


}
