% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knnmi.R
\name{cond_mutual_inf_cdd_mat}
\alias{cond_mutual_inf_cdd_mat}
\title{Conditional mutual information estimation for a continuous vector
and a discrete matrix, given another discrete matrix.}
\usage{
cond_mutual_inf_cdd_mat(x, M, Z, k = 3L, seed = 0L)
}
\arguments{
\item{x}{input vector.}

\item{M}{input matrix of the same length as x.}

\item{Z}{conditional input matrix of the same length as x.}

\item{k}{number of nearest neighbors.}

\item{seed}{integer random seed. If seed<=0 a random seed is used.}
}
\value{
a double-precision vector - mutual information estimation for
vector \code{x} and matrix \code{M}, given matrix \code{Z}.
}
\description{
Compute conditional mutual information of \code{x},\code{M} given \code{Z}
where \code{x} is continuous, \code{M} and \code{Z} are discrete
}
\examples{
data(mutual_info_df)

M <- cbind(mutual_info_df$Xd, mutual_info_df$Yd)
ZM <- cbind(mutual_info_df$Yd, mutual_info_df$Wd)
cond_mutual_inf_cdd_mat(mutual_info_df$Zc_XdYdWd, M, ZM)
## 0.1757598 0.1086227

}
