% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTNBB1_reporter.R
\docType{data}
\name{CTNBB1_reporter}
\alias{CTNBB1_reporter}
\title{A Continuous Measures of a Targeted Profile}
\format{
An object of class {\link[Biobase]{ExpressionSet}}
containing a matrix of 1000 rows (features) and 200 columns (samples) 
see (\href{http://bit.ly/2rjy29l}{ExpressionSet documentation})
}
\usage{
data(CTNBB1_reporter)
}
\description{
An actual feature matrix comprises of 17723 features where 1/0 vectors 
indicating the presence/absence of the feature in the samples. There
are 82 samples in this study.
}
\references{
Kartha VK, Kern JG, Sebastiani P, Zhang L,
Varelas X, Monti S (2017) CaDrA: A computational framework for performing 
candidate driver analyses using binary genomic features. 
(\href{https://www.biorxiv.org/content/early/2017/11/23/221846}{bioRxiv})
}
\keyword{datasets}
