% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTNBB1_transcriptional_reporter.R
\docType{data}
\name{CTNBB1_transcriptional_reporter}
\alias{CTNBB1_transcriptional_reporter}
\title{A continuous measures of a targeted profile obtained from REVEALER package}
\format{
An object of class {\link[Biobase]{ExpressionSet}}
containing a matrix of 1000 rows (features) and 200 columns (samples) 
see (\href{http://bit.ly/2rjy29l}{ExpressionSet documentation})
}
\usage{
data(CTNBB1_transcriptional_reporter)
}
\description{
Simulated matrix representing a dataset of genomic features
used as input to CaDrA for performing a stepwise search
Matrix rows comprise 1000 features represented by 1/0 vectors 
indicating the presence/absence of the feature in the sample (column)
Features include 10 left-skewed (i.e. True Positive or TP) 
and 990 uniformly-distributed (i.e. True Null or TN) 
features
}
\references{
Kartha VK, Kern JG, Sebastiani P, Zhang L,
Varelas X, Monti S (2017) CaDrA: A computational framework for performing 
candidate driver analyses using binary genomic features. 
(\href{https://www.biorxiv.org/content/early/2017/11/23/221846}{bioRxiv})
}
\keyword{datasets}
