% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revealer_functions.R
\name{kde2d_wrap}
\alias{kde2d_wrap}
\title{kde2d_wrap_ wrapper}
\usage{
kde2d_wrap(x, y, h, n = 25, lims = c(range(x), range(y)))
}
\arguments{
\item{x}{x coordinate of data}

\item{y}{y coordinate of data}

\item{h}{vector of bandwidths for x and y directions.}

\item{n}{length}

\item{lims}{The limits of the rectangle covered by the grid as c(xl, xu, yl, yu).}
}
\description{
Simplified version of Two-Dimensional Kernel Density Estimation
}
