% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knnmi_rowscore.R
\name{knnmi_rowscore}
\alias{knnmi_rowscore}
\title{knnmi Scoring Method}
\usage{
knnmi_rowscore(FS, input_score, meta_feature = NULL)
}
\arguments{
\item{FS}{a matrix of binary features where rows represent features of 
interest (e.g. genes, transcripts, exons, etc...) and columns represent 
the samples.}

\item{input_score}{a vector of continuous scores representing a phenotypic
readout of interest such as protein expression, pathway activity, etc.
The \code{input_score} object must have names or labels that match the column
names of FS object.}

\item{meta_feature}{a vector of one or more features representing known 
causes of activation or features associated with a response of interest, 
\code{e.g. input_score}. Default is NULL.}
}
\value{
return a vector of row-wise scores where its labels or names 
must match the row names of \code{FS} object
}
\description{
Compute k-Nearest Neighbor Mutual Information Estimator of \code{x} and 
\code{y} given \code{z}
}
\examples{
 
mat <- matrix(c(1,0,1,0,0,0,0,0,1,0, 
                0,0,1,0,1,0,1,0,0,0,
                0,0,0,0,1,0,1,0,1,0), nrow=3)

colnames(mat) <- 1:10
row.names(mat) <- c("TP_1", "TP_2", "TP_3")

set.seed(42)
input_score = rnorm(n = ncol(mat))
names(input_score) <- colnames(mat)

knnmi_rowscore <- knnmi_rowscore(
  FS = mat, 
  input_score = input_score,
  meta_feature = NULL
)

}
