% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_score.R
\name{ks.genescore}
\alias{ks.genescore}
\title{Directional Kolmogorov-Smirnov scoring}
\usage{
ks.genescore(n.x, y, do.pval = T, alternative = c("two.sided", "greater",
  "less"), do.plot = F, plot.dat = F, bare = F, weight = NULL,
  weight.p = 1, cls.lev = c(0, 1), absolute = F, plot.labels = FALSE,
  exact = NULL, ...)
}
\arguments{
\item{n.x}{length of ranked list}

\item{y}{positions (ranks) of geneset items in ranked list}

\item{do.pval}{a logical indicating whether or not to compute asymptotic p-value or not}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided","greater" or "less".}

\item{do.plot}{a logical indicating whether or not to draw the Enrichment Score (ES) plot}

\item{plot.dat}{a logical indicating whether or not to return the x and y axis data for the ES plot}

\item{bare}{a logical indicating whether or not to return the KS score & p-value only (a 2-tuple)}

\item{weight}{a vector of weights to use if performing a weighted-KS test. Default is NULL.}

\item{weight.p}{Exponent for weights}

\item{cls.lev}{class labels to display}

\item{absolute}{logical indicating whether or not to take the (max - min) score rather than the maximum deviation from null}

\item{plot.labels}{logical indicating whether or not to plot hits' labels}

\item{exact}{}

\item{...}{additional arguments for plot() function}
}
\value{
If plot.dat is set to TRUE, this function returns a data frame object containing x and y-axis data corresponding to the ES plot. If bare is set to TRUE, this function returns a tuple vector containing the directional ks score and p-value, respectively. If both are set to FALSE, this function returns a list object containing values corresponding to the ks statistic, p-value, alternative, method and data name
}
\description{
Compute two-sample directional KS scores
}

