% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_genescore.R
\name{ks_genescore}
\alias{ks_genescore}
\title{Compute directional KS scores for each row of a given binary matrix}
\usage{
ks_genescore(
  x,
  y,
  weight = NULL,
  weight.p = 1,
  alternative = "less",
  do.pval = TRUE,
  absolute = FALSE,
  exact = NULL
)
}
\arguments{
\item{x}{description}

\item{y}{description}

\item{weight}{description}

\item{weight.p}{description}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided","less" or "greater". Value passed to ks.genescore() function}

\item{do.pval}{description}

\item{absolute}{description}

\item{exact}{description}
}
\description{
Compute directional KS scores for each row of a given binary matrix
}
