% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cadra_functions.R
\name{ks_genescore_mat}
\alias{ks_genescore_mat}
\title{Compute KS scores for each row of a given matrix}
\usage{
ks_genescore_mat(mat, alt = "less", weight)
}
\arguments{
\item{mat}{matrix of binary features to compute row-wise ks scores for}

\item{alt}{an integer value specifying the alternative hypothesis}

\item{weight}{a vector of weights to use if performing a weighted-KS test}
}
\value{
Two lists: score and p-value
}
\description{
Compute directional Kolmogorov-Smirnov scores for each row of a 
given binary matrix
}
