% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cadra_functions.R
\name{ks_genescore_wrap}
\alias{ks_genescore_wrap}
\title{ks.genescore wrapper}
\usage{
ks_genescore_wrap(n_x, y, weight, alt = "less")
}
\arguments{
\item{n_x}{length of ranked list}

\item{y}{positions of geneset items in ranked list (ranks)}

\item{weight}{a vector of weights}

\item{alt}{alternative hypothesis for p-value calculation}
}
\value{
need a return value here
}
\description{
Compute directional Kolmogorov-Smirnov scores for each row of a given vector
}
\keyword{internal}
