% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cadra_functions.R, R/search_score.R
\name{ks_plot_wrap}
\alias{ks_plot_wrap}
\title{ks_plot wrapper}
\usage{
ks_plot_wrap(n_x, y, weight, alt = "less")

ks_plot_wrap(n_x, y, weight, alt = "less")
}
\arguments{
\item{n_x}{length of ranked list}

\item{y}{positions of geneset items in ranked list (ranks)}

\item{weight}{a vector of weights}

\item{alt}{alternative hypothesis for p-value calculation}
}
\description{
Return a dataframe from ks_genescore function

Return a dataframe from ks_genescore function
}
