% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_rowscore.R
\name{ks_rowscore}
\alias{ks_rowscore}
\title{Kolmogorov-Smirnov Scoring Method}
\usage{
ks_rowscore(
  mat,
  weights = NULL,
  alternative = c("less", "greater", "two.sided")
)
}
\arguments{
\item{mat}{a matrix of binary features to compute row-wise scores based on 
the \code{Kolmogorov-Smirnov} test.}

\item{weights}{a vector of weights to perform a \code{weighted-KS} test. 
Default is \code{NULL}. If not NULL, weights must include labels or names 
that associated with the colnames of the feature matrix.}

\item{alternative}{a character string specifies an alternative hypothesis 
testing (\code{"two.sided"} or \code{"greater"} or \code{"less"}). 
Default is \code{less} for left-skewed significance testing.}
}
\value{
A data frame with two columns: \code{score} and \code{p_value}
}
\description{
Compute directional KS scores for each row of a given binary matrix
}
\examples{

#' Load R library
library(Biobase)

#' Load pre-computed expression set
data(sim.ES)

#' Define additional parameters and start the function
result <- ks_rowscore(
  mat = exprs(sim.ES),
  weights = NULL,
  alternative = "less"
)

}
