% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cadra_functions.R, R/search_score.R
\name{ks_test_double_wrap}
\alias{ks_test_double_wrap}
\title{ks_test_d_wrap_ wrapper}
\usage{
ks_test_double_wrap(n_x, y, alt = "less")

ks_test_double_wrap(n_x, y, alt = "less")
}
\arguments{
\item{n_x}{length of ranked list}

\item{y}{positions of geneset items in ranked list (ranks)}

\item{alt}{alternative hypothesis for p-value calculation}
}
\description{
Compute directional Kolmogorov-Smirnov scores

Compute directional Kolmogorov-Smirnov scores
}
