% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/search_plot.R
\name{meta.plot}
\alias{meta.plot}
\title{Step-wise heuristic search plot}
\usage{
meta.plot(ESet, var.score = NULL, var.name = "")
}
\arguments{
\item{ESet}{an ESet object containing only the features returned from the ks.stepwise() function (if any)}

\item{var.score}{an optional integer vector of continuous measures used to rank samples for the stepwise search (assumed in matching order)}

\item{var.name}{a string object describing the name of the continuous measure used, which will be used as the y-axis label for the metric plot}
}
\value{
A plot graphic with the ranked metric plot (optional), a tile plot of the features within the provided ESet, and the corresponding Enrichment Score (ES) for a given distribution (here, this will correspond to the logical OR of the features)
}
\description{
Plot the stepwise search results for a given CaDrA run. Plot will include an optional bar plot of the continuous ranking variable (top),
}

