% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knnmi.R
\name{mutual_inf_cc_mat}
\alias{mutual_inf_cc_mat}
\title{Mutual information estimation for continuous case of an NxM matrix.}
\usage{
mutual_inf_cc_mat(x, M, k = 3L, seed = 0L)
}
\arguments{
\item{x}{input vector.}

\item{M}{input matrix. The number of rows should be equal to the
length of vector x}

\item{k}{number of nearest neighbors.}

\item{seed}{integer random seed. If seed is <=0 a random seed is used.}
}
\value{
a vector of length equal to the number of columns of matrix M.
}
\description{
Compute conditional mutual information of \code{x} and matrix \code{M}
where \code{x} and \code{M} are both continuous
}
\examples{

data(mutual_info_df)

M <- cbind(mutual_info_df$Xc, mutual_info_df$Yc)
mutual_inf_cc_mat(mutual_info_df$Zc_XcYcWc, M)
## 0.000000 0.199844

}
