% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knnmi.R
\name{mutual_inf_cc_vec}
\alias{mutual_inf_cc_vec}
\title{Mutual information estimation for continuous case of 2 vectors.}
\usage{
mutual_inf_cc_vec(x, y, k = 3L, seed = 0L)
}
\arguments{
\item{x}{input vector.}

\item{y}{input vector of the same length as x.}

\item{k}{number of nearest neighbors.}

\item{seed}{integer random seed. If seed is <=0 a random seed is used.}
}
\value{
a double-precision value - mutual information estimation for
vectors \code{x} and \code{y}.
}
\description{
Compute mutual information of \code{x} and \code{y}
where \code{x} and \code{y} are both continuous
}
\examples{

data(mutual_info_df)
mutual_inf_cc_vec(mutual_info_df$Xc, mutual_info_df$Zc_XcYc)
## 0

mutual_inf_cc_vec(mutual_info_df$Yc, mutual_info_df$Zc_XcYc)
## 0.2738658

}
