% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knnmi.R
\name{mutual_inf_cd_vec}
\alias{mutual_inf_cd_vec}
\title{Mutual information estimation for continuous/discrete case.}
\usage{
mutual_inf_cd_vec(x, y, k = 3L, seed = 0L, use_cc = FALSE)
}
\arguments{
\item{x}{input (continuous) vector.}

\item{y}{input (discrete) vector. It should have the same length as x.}

\item{k}{number of nearest neighbors.}

\item{seed}{integer random seed. If seed<=0 a random seed is used.}

\item{use_cc}{(logical) if TRUE the algorithm falls into continuous/continuous case}
}
\value{
a double-precision value - mutual information estimation for
vectors \code{x} and \code{y}.
}
\description{
Compute mutual information of \code{x} and \code{y}
where \code{x} is continuous vector and \code{y} are discrete
}
\examples{

data(mutual_info_df)

mutual_inf_cd_vec(mutual_info_df$Zc_XdYd, mutual_info_df$Xd)
## 0.128029

}
