% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mutual_info_df}
\alias{mutual_info_df}
\title{Toy Dataset for knnmi package}
\format{
A data frame with 100 rows and 10 columns
}
\usage{
data(mutual_info_df)
}
\description{
Toy Dataset for knnmi package
}
\keyword{datasets}
