% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{null.search}
\alias{null.search}
\title{Permutation-based step-wise searching}
\usage{
null.search(
  ranking = NULL,
  ES,
  nperm = 1000,
  plot = TRUE,
  obs.best.score = NULL,
  cache.path = NULL,
  smooth = TRUE,
  return.perm.pval = FALSE,
  seed = 123,
  ncores = 1,
  ...
)
}
\arguments{
\item{ranking}{a vector containing rankings for sample ordering. Default is NULL. If NULL, we assume the ESet is already ranked. This is used only to compute the observed stepwise search score, if not provided}

\item{ES}{an ordered expression set object with the same sample ordering and features as processed by the stepwise.search() function when performing a step-wise heuristic search}

\item{nperm}{an integer specifying the number of permutations to perform. Default = 1000}

\item{plot}{logical indicating whether or not to plot the emperical null distribution with the observed score and permutation p-value}

\item{obs.best.score}{a numeric value corresponding to the observed (best) stepwise search score to use for permutation based p-value computation. Default is NULL. If set to NULL, we compute the observed score given the ranking variable and ESet}

\item{cache.path}{full path to permutation-based (null) score distributions cache files. If a permutation for a given dataset (and dependent search variables such as 'N') exist, we recycle values instead of re-computing them to save time. Default is NULL. If NULL, cache path is set to the default ~/.Rcache for future cache loading.}

\item{smooth}{logical indicating whether or not to smoothen the p-value calculation to avoid p-value of 0. Default is TRUE}

\item{return.perm.pval}{logical indicating whether or not to return the permutation-based p-value computed by the function. Default is FALSE}

\item{seed}{seed set for permutation. Default = 123}

\item{ncores}{number of cores to use, if using parallelization for permutation testing. Default = 1}

\item{...}{additional parameters passed to the stepwise.seach() function, which will be applied to each permutation run (called within topn.eval())}
}
\value{
If return.perm.pval is set to TRUE, will return the permutation p-value
}
\description{
Performs permutation-based significance testing of step-wise search results.
}
