% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwise_search.R
\name{null_ks}
\alias{null_ks}
\title{Permutation-based step-wise searching}
\usage{
null_ks(ranking = NULL, ES, metric, nperm = 1000, back_search = T,
  cust_start = NULL, plot = T, alt = "two.sided", wts = NULL,
  smooth = T, seed = 123, ncores = 1)
}
\arguments{
\item{ranking}{a vector containing rankings for sample ordering. Default is NULL. If NULL, we assume the ESet is already ranked. This is used only to compute the observed KS stepwise search score}

\item{ES}{an ordered expression set object with the same sample ordering and features as processed by the ks.stepwise() function when performing a step-wise heuristic search}

\item{metric}{a character string specifying which metric to use for stepwise search criteria. One of either 'pval' or 'stat' may be used, corresponding to the KS p-value or statistic}

\item{nperm}{an integer specifying the number of permutations to perform. Default = 1000}

\item{back_search}{a logical indicating whether or not to perform a forward-backward search (i.e. remove features along the search if it improves the best score). Default is TRUE. Uses function backward_check()}

\item{cust_start}{an integer specifying a specific index within the expression set object of the feature to start the step-wise search with. Default is NULL}

\item{plot}{logical indicating whether or not to plot the emperical null distribution with the observed p-value}

\item{alt}{a character string specifying the alternative hypothesis, must be one of "two.sided","greater" or "less". Default is "two.sided"}

\item{wts}{a vector of weights to use if performing weighted-KS testing. Default is NULL}

\item{smooth}{logical indicating whether or not to smoothen the p-value calculation to avoid p-value of 0. Default is TRUE}

\item{seed}{seed set for permutation. Default = 123}

\item{ncores}{number of cores to use, if using parallelization. Default = 1}
}
\value{
a list object containing the permutation-based p-value based on the generated empirical null distribution of stepwise KS searching
}
\description{
Performs permutation-based significance testing of step-wise KS search.
}

