% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm_data.R
\docType{data}
\name{perm.res}
\alias{perm.res}
\title{Permutation Results}
\format{
A list of objects returned by \code{cadra_search()}
containing parameters such as top N features, search start, permutation best scores, permutation p-values and observed best score. 
This resulting object can be used to visualize the null permutation distribution plot.
See \code{\link[CaDrA]{cadra_search}} for more information.
}
\usage{
data(perm.res)
}
\description{
A list of objects returned from cadra_search() when it runs
on the \code{sim.ES} simulated dataset with ks method.
}
\examples{
data(perm.res)
}
\references{
Kartha VK, Kern JG, Sebastiani P, Zhang L,
Varelas X, Monti S (2017) CaDrA: A computational framework for performing 
candidate driver analyses using binary genomic features. 
(\href{https://www.biorxiv.org/content/early/2017/11/23/221846}{bioRxiv})
}
\keyword{datasets}
