% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_plot.R
\name{plot_comb}
\alias{plot_comb}
\title{Step-wise heuristic search plot}
\usage{
plot_comb(ESet, d, ES.dat, variable_name = "", row_std = F)
}
\arguments{
\item{ESet}{an ESet object containing only the features returned from the ks.stepwise() function (if any)}

\item{d}{a data frame object with columns 'sample' and 'measure' containing the continuous measures used to rank samples for the stepwise search (in matching order)}

\item{ES.dat}{a data frame object with columns 'x' and 'y' containing the necessary data for an ES plot returned by the ks.genescore() function when plot.dat is set to TRUE for the given dataset}

\item{variable_name}{a string object describing the name of the continuous measure used, which will be used as the y-axis label for the metric plot}

\item{row_std}{logical indicating whether or not to perform row Z-score transformation on the data before visualizing (only for continuous data). Default is FALSE}
}
\value{
A plot graphic with the ranked metric plot, a tile plot of the features within the provided ESet, and the corresponding Enrichment Score (ES) for a given distribution (here, this will correspond to the union of the features, which we assume is the last row in the ESet feature space)
}
\description{
Plot together the continuous variable used to rank samples in ks.stepwise()
}

