% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_plot.R
\name{plot_enrichment_plot}
\alias{plot_enrichment_plot}
\title{Enrichment Score Plot}
\usage{
plot_enrichment_plot(df)
}
\arguments{
\item{df}{a data frame object with columns 'X' and 'Y' containing the
x, y coordinates for the enrichment plot}
}
\value{
A graphic of Enrichment Scores for a given distribution
}
\description{
An enrichment plot of a sum statistic across a given meta-features. The x and y-axis
data are returned from ks_plot_coordinates()
}
\examples{

# Load R library
library(SummarizedExperiment)

# Load pre-computed Top-N list generated for sim_FS dataset
data(topn_list)

# With the results obtained from top-N evaluation,
# We can find the combination of features that gives
# the best score in top N searches
topn_best_meta <- topn_best(topn_list=topn_list)

# Extract the meta-feature set
feature_set <-  topn_best_meta[["feature_set"]]

# Make sure mat variable is a matrix
mat <- as.matrix(assay(feature_set))

# For cases when matrix only has one row
if(ncol(mat) == 1){
  mat <- matrix(t(mat), nrow=1, byrow=TRUE,
  dimnames = list(rownames(feature_set), rownames(mat)))
}

# Add on the OR function of all the returned entries
or <- ifelse(colSums(mat)==0, 0, 1)

# Get x and y axis data for enrichment plot of
# cumulative function of individual features
# (i.e. the OR function)
enrichment_dat <- ks_plot_coordinates(
   n.x = length(or),
   y = which(or==1),
   weight = NULL,
   alt = "less"
)

# Plot the enrichment scores
plot_enrichment_plot(df = enrichment_dat)

}
