% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_plot.R
\name{plot_ESet}
\alias{plot_ESet}
\title{Enrichment Score (ES) plot}
\usage{
plot_ESet(d)
}
\arguments{
\item{d}{a data frame object with columns 'x' and 'y' containing the necessary data for an ES plot (returned by ks.genescore() when plot.dat is set to TRUE for a given dataset)}
}
\value{
A plot graphic of the Enrichment Score (ES) for a given distribution
}
\description{
Plot the ES running sum statistic, including the max score (KS) using x and y-axis data returned by ks.genescore() when plot.dat is set to TRUE for a given dataset
}
