% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwise_search.R
\name{prefilter_data}
\alias{prefilter_data}
\title{Pre-filter features}
\usage{
prefilter_data(ES, cutoff = 0.6)
}
\arguments{
\item{ES}{an expression set object containing binary features used for step-wise search}

\item{cutoff}{a fraction between 0 and 1 describing the absolute prevalence of a feature across all samples in the dataset above which the feature will be filtered out. Default is 0.6 (feature should occur in 60 percent or more of the samples to be removed)}
}
\value{
An expression set object with the filtered-in features only
}
\description{
Pre-filter a dataset prior to running step-wise heuristic search in order to avoid testing features that are too prevalent across samples in the dataset
}

