% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_plot.R
\name{rbind_gtable_max}
\alias{rbind_gtable_max}
\title{Adjusted row-bind of ggplotGrob objects}
\usage{
rbind_gtable_max(...)
}
\arguments{
\item{...}{multiple ggplotGrob objects using which a list is made}
}
\value{
a gtable object with the width-adjusted stack of ggplots
}
\description{
Row-bind ggplotGrob objects so that multiple, stacked plots are aligned based on the maximum width of any plots. in the table. Each plot will be left-aligned. Each plot should be a ggplotGrob object. Ex: grid.draw(rbind_gtable_max(ggplotGrob(ggplot1),ggpotGrob(ggplot2)))
}

