% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revealer_rowscore.R
\name{revealer_score}
\alias{revealer_score}
\title{Compute Conditional Mutual Information of x and y given z from \code{REVEALER}}
\usage{
revealer_score(x, y, z = NULL, assoc_metric = c("IC", "COR"))
}
\arguments{
\item{x}{a vector of continuous values of
a given functional response of interest}

\item{y}{a binary present/absent feature typically representing
genome-wide alterations (mutations, cnvs, amplifications/deletions)}

\item{z}{a consolidated or summarized vector of values obtained from
one or more binary features(s) which representing known “causes”
of activation or features associated with a given response of interest}

\item{assoc_metric}{an association metric: information coefficient
(\code{"IC"} by default) or correlation (\code{"COR"}) from \code{REVEALER}.}
}
\value{
a data frame with two columns: \code{score} and \code{p_value}
}
\description{
Compute Conditional Mutual Information of x and y given z from \code{REVEALER}
}
