% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_plot.R
\name{row_standard}
\alias{row_standard}
\title{Standardize matrix by row}
\usage{
row_standard(mat)
}
\arguments{
\item{mat}{a matrix object of continuous values to be row-standardized}
}
\value{
A Z-score transformed matrix with the same dimensions as the input matrix
}
\description{
Standardize a matrix by row, converting the values to a Z score
}

