% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_scores.R
\docType{data}
\name{sim.Scores}
\alias{sim.Scores}
\title{Simulated Input Scores}
\format{
A vector of continuous scores generated from rnorm(n=ncol(sim.ES), 
mean=0, sd=1) with seed=123
}
\usage{
data(sim.Scores)
}
\description{
A simulated vector of continuous scores generated from rnorm(n=ncol(sim.ES), 
mean=0, sd=1) with seed=123 based on the number of samples in simulated 
Expressionset (sim.ES)
}
\references{
Kartha VK, Kern JG, Sebastiani P, Zhang L,
Varelas X, Monti S (2019) CaDrA: A computational framework for performing 
candidate driver analyses using binary genomic features. 
(\href{https://www.frontiersin.org/articles/10.3389/fgene.2019.00121/full}{Frontiers in Genetics})
}
\keyword{datasets}
