% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{topn.best}
\alias{topn.best}
\title{Top 'N' best}
\usage{
topn.best(topn.list)
}
\arguments{
\item{topn.list}{The nested list object that is returned by topn.eval() function when best.score.only is set to FALSE. This contains both the ESets as well as the scores for each resulting meta-feature in the top 'N'  search mode.}
}
\value{
A list containing the (local) best meta-feature ESet, as well as its corresponding search score
}
\description{
Takes the resulting list of meta-features returned from topn.eval() function and fetches the meta-feature with best (local) score
}
