% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwise_search.R
\name{topn.eval}
\alias{topn.eval}
\title{Top 'N' evaluate}
\usage{
topn.eval(l, do.plot = TRUE)
}
\arguments{
\item{l}{a list of expression sets returned by each stepwise search run of the top 'n' specified features}

\item{do.plot}{a logical indicating whether you want to only plot the resulting evaluation matrix instead of returning it. Default is TRUE}
}
\value{
A binary overlap matrix (or a heatmap thereof) with rows being the union of all reported features and the column being the starting feature for each run. 
1's and 0's represent whether a feature in any given row is present in a meta-feature along with a starting feature in the corresponding column.
}
\description{
Evaluates a list of stepwise search results from the ks.stepwise function for overlapping features. This function is mainly used to evaluate search results over the top 'N' best starting features for a given dataset.
}

