% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topn_best.R
\name{topn_best}
\alias{topn_best}
\title{Top 'N' best meta-features}
\usage{
topn_best(topn_list)
}
\arguments{
\item{topn_list}{A nested list object that is returned by topn_eval() 
function when best_score_only is set to FALSE. This contains both the 
ESets as well as the scores and its input scores for each resulting 
meta-feature in the top 'N' search mode.}
}
\value{
A list containing the (local) best meta-feature ESet, as well as 
its corresponding search score and input scores
}
\description{
Takes the resulting list of meta-features returned from topn_eval() 
function and fetches the meta-features with best (local) score
}
\examples{

# Load pre-computed Top-N list generated for sim.ES dataset
data(topn.list)

# Get the best meta-features list
topn_best_meta <- topn_best(topn_list=topn.list) 

}
