% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topn_eval.R
\name{topn_eval}
\alias{topn_eval}
\title{Top 'N' evaluate}
\usage{
topn_eval(
  ESet,
  input_score,
  method = "ks",
  alternative = "less",
  metric = "pval",
  search_method = "both",
  max_size = 7,
  best_score_only = TRUE,
  N = 1,
  do_plot = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{ESet}{an ordered expression set object with the same sample ordering and features as processed by the stepwise.search() function when performing a step-wise heuristic search}

\item{input_score}{a vector of ranked or continuous values (required).}

\item{method}{a character string specifying the method used to compute scores for features, must be one of "ks" or "wilcox" or "mi" (mutually exclusive method from REVEALER) or "custom" (a personal customization method). If input_score contains ranked scores, then 'ks' method is used by default. Otherwise, 'mi" is the default method}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided", "greater" or "less". Default is "less" for left-skewed significance testing.}

\item{metric}{a character string specifying which metric to use for candidate search. One of either 'pval' or 'stat' may be used, corresponding to the score p-value or statistic. Default is 'pval'}

\item{search_method}{a character string specifying which method to perform or filter out the best candidates. Default is 'forward'.}

\item{max_size}{an integer specifying the maximum size a meta-feature can extend do for a given search. Default is 7.}

\item{best_score_only}{a logical indicating whether or not to only return the best meta-feature score over the top 'N' evaluation. Default is TRUE}

\item{N}{an integer specifying the number of features to start the search over, starting from the top 'N' features in each case. Default is 1}

\item{do_plot}{a logical indicating whether you want to plot the resulting evaluation matrix. Default is TRUE}

\item{verbose}{a logical indicating whether or not to verbose diagnostic messages. Default is TRUE.}
}
\value{
Default is a list of lists, where each list entry is one that is returned by the stepwise search run for a given starting index (See stepwise.search()). If best_score_only is set to TRUE, only the best score over the top N space is returned (useful for permutation-based testing)
1's and 0's represent whether a feature in any given row is present in a meta-feature along with a starting feature in the corresponding column.
}
\description{
Generates and evaluates stepwise search results for the top 'N' starting indices, checking for overlapping resulting features from each case. This function is mainly used to evaluate search results over the top 'N' best starting features for a given dataset.
}
