% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_score.R
\name{wilcox.genescore.mat}
\alias{wilcox.genescore.mat}
\title{Row-wise matrix Wilcoxon rank sum scoring}
\usage{
wilcox.genescore.mat(
  mat,
  alt = c("two.sided", "less", "greater"),
  ranks = NULL
)
}
\arguments{
\item{mat}{matrix of binary features to compute row-wise scores for based on the Wilcoxon rank sum test}

\item{alt}{a character string specifying the alternative hypothesis, must be one of "two.sided","less" or "greater". Value passed to wilcox.genescore() function}

\item{ranks}{a vector of ranks to use when performing the Wilcoxon test. Default is NULL. If NULL, then samples are assumed to be ordered by increasing ranking. Value passed to wilcox.genescore() function}
}
\value{
A 2 x N matrix (N = number of rows in matrix) with the first and second rows corresponding to the rank sum statistic scores and p-values, respectively
}
\description{
Compute rank sum scores for each row of a given binary matrix
}
