% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcox_genescore.R
\name{wilcox_genescore}
\alias{wilcox_genescore}
\title{Compute rank sum scores for each row of a given binary matrix}
\usage{
wilcox_genescore(
  x,
  y,
  mu = 0,
  alternative = "less",
  paired = FALSE,
  exact = FALSE,
  correct = TRUE
)
}
\arguments{
\item{x}{description}

\item{y}{description}

\item{mu}{description}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided","less" or "greater". Value passed to ks.genescore() function}

\item{paired}{description}

\item{exact}{description}

\item{correct}{description}
}
\description{
Compute rank sum scores for each row of a given binary matrix
}
