% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcox_genescore_mat.R
\name{wilcox_genescore_mat}
\alias{wilcox_genescore_mat}
\title{Wilcoxon Rank Sum Scoring Method}
\usage{
wilcox_genescore_mat(
  mat,
  ranks = NULL,
  alternative = c("less", "greater", "two.sided"),
  verbose = FALSE
)
}
\arguments{
\item{mat}{a matrix of binary features (required).}

\item{ranks}{a vector of sample rankings use to perform the 
\code{Wilcoxon test}. Default is \code{NULL}. If NULL, 
then samples are assumed to be ordered by increasing rankings. 
If not NULL, ranks must include labels or names that associated 
with the colnames of the feature matrix.}

\item{alternative}{a character string specifies an alternative 
hypothesis testing (\code{"two.sided"} or \code{"greater"} or 
\code{"less"}). Default is \code{less} for left-skewed significance testing.}

\item{verbose}{a logical value indicates whether or not to 
print the diagnostic messages. Default is \code{FALSE}.}
}
\value{
A data frame with two columns: \code{score} and \code{p_value}
}
\description{
Compute directional Wilcoxon rank sum score for each row of a 
given binary matrix
}
\examples{

# Load R library
library(Biobase)
# Load pre-computed expression set
data(sim.ES)

# Define additional parameters and run the function
wilcox_genescore_mat_result <- wilcox_genescore_mat(
  mat = exprs(sim.ES), 
  ranks = NULL,
  alternative = "less"
)

}
