% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcox_rowscore.R
\name{wilcox_rowscore}
\alias{wilcox_rowscore}
\title{Wilcoxon Rank Sum Scoring Method}
\usage{
wilcox_rowscore(
  mat,
  ranks = NULL,
  alternative = c("less", "greater", "two.sided")
)
}
\arguments{
\item{mat}{a matrix of binary features (required).}

\item{ranks}{a vector of sample rankings use to perform the 
\code{Wilcoxon test}. Default is \code{NULL}. If NULL, 
then samples are assumed to be ordered by increasing rankings. 
If not NULL, ranks must include labels or names that associated 
with the colnames of the feature matrix.}

\item{alternative}{a character string specifies an alternative 
hypothesis testing (\code{"two.sided"} or \code{"greater"} or 
\code{"less"}). Default is \code{less} for left-skewed significance testing.}
}
\value{
A data frame with two columns: \code{score} and \code{p_value}
}
\description{
Compute directional Wilcoxon rank sum score for each row of a 
given binary matrix
}
\examples{

# Load R library
library(Biobase)
# Load pre-computed expression set
data(sim.ES)

# Define additional parameters and run the function
result <- wilcox_rowscore(
  mat = exprs(sim.ES), 
  ranks = NULL,
  alternative = "less"
)

}
