% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRCA_GISTIC_MUT_SIG.R
\docType{data}
\name{BRCA_GISTIC_MUT_SIG}
\alias{BRCA_GISTIC_MUT_SIG}
\title{Genomic Data from the TCGA BRCA}
\format{
An object of class \code{ExpressionSet} from \code{Biobase} package
containing a matrix of 16873 rows (features) and 963 columns (samples) 
see (\href{http://bit.ly/2rjy29l}{ExpressionSet documentation})
}
\usage{
data(BRCA_GISTIC_MUT_SIG)
}
\description{
Binary matrix of mutation and SCNA features comprises of 16873 features where 1/0 vectors 
indicating the presence/absence of the feature in the samples. There
are 963 samples in this cohort study.
}
\references{
Kartha VK, Kern JG, Sebastiani P, Zhang L,
Varelas X, Monti S (2017) CaDrA: A computational framework for performing 
candidate driver analyses using binary genomic features. 
(\href{https://www.biorxiv.org/content/early/2017/11/23/221846}{bioRxiv})
}
\keyword{datasets}
