\name{CaDrA-package}
\alias{CaDrA-package}
\alias{CaDrA}
\docType{package}
\title{
Candidate Driver Analysis
}
\description{
This is a package to support analyses involving the identification of candidate upstream (epi) genetic regulators of pathway activity based on a step-wise heuristic search algorithm.
}
\details{
\tabular{ll}{
Package: \tab CaDrA\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-11-24\cr
License: \tab What license is it under?\cr
}
library(CaDrA)
}
\author{
Vinay Kartha
Stefano Monti

Maintainer: Vinay Kartha <vkartha@bu.edu>
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
