% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cadra.R
\name{CaDrA}
\alias{CaDrA}
\title{CaDrA Search}
\usage{
CaDrA(
  FS,
  input_score,
  method = c("ks_pval", "ks_score", "wilcox_pval", "wilcox_score", "revealer", "custom"),
  custom_function = NULL,
  custom_parameters = NULL,
  alternative = c("less", "greater", "two.sided"),
  weight = NULL,
  top_N = 1,
  search_start = NULL,
  search_method = c("both", "forward"),
  max_size = 7,
  n_perm = 1000,
  obs_best_score = NULL,
  smooth = TRUE,
  plot = TRUE,
  ncores = 1,
  cache_path = NULL,
  verbose = FALSE
)
}
\arguments{
\item{FS}{a SummarizedExperiment object containing binary features where
rows represent features of interest (e.g. genes, transcripts, exons, etc...)
and columns represent the samples.}

\item{input_score}{a vector of continuous scores representing a phenotypic
readout of interest such as protein expression, pathway activity, etc.

NOTE: \code{input_score} object must have names or labels that match the column
names of \code{FS} object.}

\item{method}{a character string specifies a scoring method that is
used in the search. There are 6 options: (\code{"ks_pval"} or \code{ks_score}
or \code{"wilcox_pval"} or \code{wilcox_score} or 
\code{"revealer"} (conditional mutual information from REVEALER) or
\code{"custom"} (a customized scoring method)). 
Default is \code{ks_pval}.}

\item{custom_function}{if method is \code{"custom"}, specifies
the name of the customized function here. Default is \code{NULL}.

NOTE: custom_function() must take \code{FS} and \code{input_score} 
as its input arguments, and its final result must return a vector of row-wise 
scores ordered from most significant to least significant where its labels or 
names matched the row names of \code{FS} object.}

\item{custom_parameters}{if method is \code{"custom"}, specifies a list of
additional arguments (excluding \code{FS} and \code{input_score}) to be 
passed to \code{custom_function}. Default is \code{NULL}.}

\item{alternative}{a character string specifies an alternative hypothesis
testing (\code{"two.sided"} or \code{"greater"} or \code{"less"}).
Default is \code{less} for left-skewed significance testing.
NOTE: this argument only apply to KS and Wilcoxon method}

\item{weight}{if method is \code{ks_score}, specifies a vector of weights 
to perform a weighted-KS testing. Default is \code{NULL}.}

\item{top_N}{an integer specifies the number of features to start the
search over. By default, it starts from the top best feature (top_N = 1).
NOTE: If \code{top_N} is provided, then \code{search_start} parameter
will be ignored.}

\item{search_start}{a list of character strings (separated by commas)
which specifies feature names within the FS object to start
the search with. If \code{search_start} is provided, then \code{top_N}
parameter will be ignored. Default is \code{NULL}.}

\item{search_method}{a character string specifies an algorithm to filter out
the best candidates (\code{"forward"} or \code{"both"}). Default is
\code{both} (i.e., backward and forward).}

\item{max_size}{an integer specifies a maximum size that a meta-feature can
extend to do for a given search. Default is \code{7}.}

\item{n_perm}{an integer specifies the number of permutations to perform.
Default is \code{1000}.}

\item{obs_best_score}{a numeric value corresponding to the best observed
score. This value is used to compare against the permuted best scores.
Default is \code{NULL}. If set to NULL, we will compute the observed
best score based on the given parameters.}

\item{smooth}{a logical value indicates whether or not to smooth the p-value
calculation to avoid p-value of 0. Default is \code{TRUE}.}

\item{plot}{a logical value indicates whether or not to plot the empirical
null distribution of the permuted best scores. Default is \code{TRUE}.}

\item{ncores}{an integer specifies the number of cores to perform
parallelization for permutation-based testing. Default is \code{1}.}

\item{cache_path}{a full path uses to cache the permuted best scores.
We recycle these scores instead of re-computing them to save time.
Default is \code{NULL}. If NULL, the cache path is set to \code{~/.Rcache}
for future loading.}

\item{verbose}{a logical value indicates whether or not to print the
diagnostic messages. Default is \code{FALSE}.}
}
\value{
a list of key parameters that are used to cache the result of
permutation-based testing, a vector of permuted best scores for a given
\code{n_perm}, an observed best score, and a permutation p-value.
}
\description{
Perform permutation-based testings on a sample of permuted input scores 
using \code{candidate_search} as the main iterative function for each run.
}
\examples{

# Load pre-computed feature set
data(sim_FS)

# Load pre-computed input-score
data(sim_Scores)

# Define additional parameters and start the function
cadra_result <- CaDrA(
  FS = sim_FS, input_score = sim_Scores, method = "ks_pval", 
  weight = NULL, alternative = "less", top_N = 1,
  search_start = NULL, search_method = "both", max_size = 7, 
  n_perm = 10, plot = FALSE, smooth = TRUE, obs_best_score = NULL,
  ncores = 1, cache_path = NULL
)

}
\author{
Reina Chau
}
