% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{CaDrA_Server}
\alias{CaDrA_Server}
\title{Shiny Server modules}
\usage{
CaDrA_Server(id)
}
\arguments{
\item{id}{A unique namespace identifier}
}
\value{
Shiny server elements
}
\description{
Shiny Server modules
}
