% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rowscore.R
\name{calc_rowscore}
\alias{calc_rowscore}
\title{Calculate row-wise scores of a given binary feature set based on
a given scoring method}
\usage{
calc_rowscore(
  FS_mat,
  input_score,
  method = c("ks_pval", "ks_score", "wilcox_pval", "wilcox_score", "revealer",
    "custom_pval", "custom_score"),
  alternative = c("less", "greater", "two.sided"),
  weight,
  seed_names = NULL,
  custom_function,
  custom_parameters,
  do_check = TRUE
)
}
\arguments{
\item{FS_mat}{a matrix of binary features where 
rows represent features of interest (e.g. genes, transcripts, exons, etc...)
and columns represent the samples.}

\item{input_score}{a vector of continuous scores representing a phenotypic
readout of interest such as protein expression, pathway activity, etc.
The \code{input_score} object must have names or labels that match the column
names of FS object.}

\item{method}{a character string specifies a scoring method that is
used in the search. There are 7 options: (\code{"ks_pval"} or \code{ks_score}
or \code{"wilcox_pval"} or \code{wilcox_score} or 
\code{"revealer"} (conditional mutual information from REVEALER) or
\code{"custom_pval"} or \code{custom_score} (a user customized scoring method)). 
Default is \code{ks_pval}.}

\item{alternative}{a character string specifies an alternative hypothesis
testing (\code{"two.sided"} or \code{"greater"} or \code{"less"}).
Default is \code{less} for left-skewed significance testing.
NOTE: \code{Revealer} method only returns score statistics (no p-values).}

\item{weight}{if method is \code{ks}, specifies a vector of weights
to perform a weighted-KS testing. Default is \code{NULL}.}

\item{seed_names}{one or more features representing known “causes”
of activation or features associated with a response of interest.
It is used for \code{method = revealer} only.}

\item{custom_function}{if method is \code{"custom"}, just pass the name of the
customized function. Default is \code{NULL}.}

\item{custom_parameters}{if method is \code{"custom"}, specifies a list of
additional arguments (other than \code{FS} and \code{input_score}) to be passed
to the \code{custom_function()}. Default is \code{NULL}.}

\item{do_check}{a logical value indicates whether or not to check
diagnostic messages. Default is \code{TRUE}}
}
\value{
a matrix with one column (score or p_value) containing row-wise 
directional scores based on a given scoring method
}
\description{
Calculate row-wise scores of a given binary feature set based on
a given scoring method
}
\examples{

# Load library
library(SummarizedExperiment)

# Load pre-computed feature set
data(sim_FS)

# Load pre-computed input-score
data(sim_Scores)

# Run the ks method
ks_rowscore_result <- calc_rowscore(
  FS_mat = assay(sim_FS),
  input_score = sim_Scores,
  method = "ks_pval",
  weight = NULL,
  alternative = "less"
)

# Run the wilcoxon method
wilcox_rowscore_result <- calc_rowscore(
  FS_mat = assay(sim_FS),
  input_score = sim_Scores,
  method = "wilcox_pval",
  alternative = "less"
)

# Run the revealer method
revealer_rowscore_result <- calc_rowscore(
  FS_mat = assay(sim_FS),
  input_score = sim_Scores,
  method = "revealer",
  seed_names = NULL
)

}
