% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCLE_MUT_SCNA.R
\docType{data}
\name{CCLE_MUT_SCNA}
\alias{CCLE_MUT_SCNA}
\title{Genomic Data from CCLE MUT + SCNA}
\format{
An object of class \code{SummarizedExperiment} from \code{SummarizedExperiment} package
containing a matrix of 17,723 rows (features) and 82 columns (samples).
See \code{?SummarizedExperiment} for more details.
}
\usage{
data(CCLE_MUT_SCNA)
}
\description{
A SummarizedExperiment object consists of 17,723 genomic features across 82 samples.
}
\references{
Kim, J., Botvinnik, O., Abudayyeh, O. et al.
Characterizing genomic alterations in cancer by complementary functional
associations. Nat Biotechnol 34, 539–546 (2016).
\href{https://doi.org/10.1038/nbt.3527}{https://doi.org/10.1038/nbt.3527}
}
\keyword{datasets}
