% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cadra.R
\name{generate_permutations}
\alias{generate_permutations}
\title{Random permutation matrix generator}
\usage{
generate_permutations(ord, n_perms, verbose = FALSE)
}
\arguments{
\item{ord}{vector to be permuted. This determines the number of columns in the permutation matrix}

\item{n_perms}{number of permutations to generate. This determines the number of rows in the permutation matrix}

\item{verbose}{a logical value indicates whether or not to print the diagnostic messages. Default is \code{FALSE}.}
}
\value{
A row matrix of permuted values (i.e. ranks) where each row is a single permutation result
}
\description{
Produces a random permutation rank matrix given a vector of values
}
\examples{

# Load pre-computed input score
data(TAZYAP_BRCA_ACTIVITY)
input_score = TAZYAP_BRCA_ACTIVITY

# Set seed
set.seed(123)

# Define number of permutations
n_perm = 1000
 
# Define additional parameters and start the function
perm_labels_matrix <- generate_permutations(ord=input_score, n_perms=n_perm)
  
}
