% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cadra_search.R
\name{generate_permutations}
\alias{generate_permutations}
\title{Random permutation matrix generator}
\usage{
generate_permutations(ord, n_perms, seed = 123)
}
\arguments{
\item{ord}{vector to be permuted. This determines the number of columns in the permutation matrix}

\item{n_perms}{number of permutations to generate. This determines the number of rows in the permutation matrix}

\item{seed}{seed which can be set for reproducibility of 'random' results. Default is 123}
}
\value{
A row matrix of permuted values (i.e. ranks) where each row is a single permutation result
}
\description{
Produces a random permutation rank matrix given a vector of values
}
