% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/search_score.R
\name{ks.genescore.mat}
\alias{ks.genescore.mat}
\title{Row-wise matrix Kolmogorov-Smirnov scoring}
\usage{
ks.genescore.mat(mat, alt = c("two.sided", "less", "greater"),
  weight = NULL)
}
\arguments{
\item{mat}{matrix of binary features to compute row-wise scores for based on the  Kolmogorov-Smirnov test}

\item{alt}{a character string specifying the alternative hypothesis, must be one of "two.sided","less" or "greater". Value passed to ks.genescore() function}

\item{weight}{a vector of weights to use if performing a weighted-KS test. Default is NULL. Value passed to ks.genescore() function}
}
\value{
A 2 x N matrix (N = number of rows in matrix) with the first and second rows corresponding to the KS scores and p-values, respectively
}
\description{
Compute directional KS scores for each row of a given binary matrix
}

