% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_gene_score.R
\name{ks_gene_score}
\alias{ks_gene_score}
\title{Compute a Kolmogorov-Smirnov score for a given ranked list}
\usage{
ks_gene_score(
  x,
  y,
  weights = NULL,
  weight_p = 1,
  alternative = "less",
  do_pval = TRUE,
  absolute = FALSE,
  exact = NULL
)
}
\arguments{
\item{x}{ranked list}

\item{y}{positions of geneset items in ranked list (ranks)}

\item{weights}{weights for weighted score (see Subramanian et al.) (usually, sort(score))}

\item{weight_p}{weights exponent}

\item{alternative}{alternative hypothesis for p-value calculation}

\item{do_pval}{compute asymptotic p-value}

\item{absolute}{takes max - min score rather than the maximum deviation from null}

\item{exact}{compute exact p-value}
}
\value{
a data frame with two columns: \code{score} and \code{p_value}
}
\description{
Compute a Kolmogorov-Smirnov score for a given ranked list
}
