% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_gene_score_mat.R
\name{ks_gene_score_mat}
\alias{ks_gene_score_mat}
\title{Row-wise matrix Kolmogorov-Smirnov scoring}
\usage{
ks_gene_score_mat(mat, weight = NULL, alternative = "less")
}
\arguments{
\item{mat}{matrix of binary features to compute row-wise scores for based on the Kolmogorov-Smirnov test}

\item{weight}{a vector of weights to use if performing a weighted-KS test. Default is NULL. Value passed to ks_genescore() function}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided","less" or "greater". Value passed to ks_genescore() function}
}
\value{
A data frame
}
\description{
Compute directional KS scores for each row of a given binary matrix
}
