% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_genescore.R
\name{ks_genescore}
\alias{ks_genescore}
\title{Compute directional KS scores for each row of a given binary matrix}
\usage{
ks_genescore(
  x,
  y,
  weight = NULL,
  weight_p = 1,
  alternative = "less",
  do_pval = TRUE,
  absolute = FALSE,
  exact = NULL
)
}
\arguments{
\item{x}{ranked list}

\item{y}{positions of geneset items in ranked list (ranks)}

\item{weight}{weights for weighted score (see Subramanian et al.) (usually, sort(score))}

\item{weight_p}{weights exponent}

\item{alternative}{alternative hypothesis for p-value calculation}

\item{do_pval}{compute asymptotic p-value}

\item{absolute}{takes max - min score rather than the maximum deviation from null}

\item{exact}{compute exact p-value}
}
\value{
A data frame
}
\description{
Compute directional KS scores for each row of a given binary matrix
}
