% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm_data.R
\docType{data}
\name{perm.res}
\alias{perm.res}
\title{Permutation Results}
\format{
A list of objects returned from \code{CaDrA()}
containing top_N value, search_start value, permutation best scores across 
n_perm, permutation p-value and observed best score. 
The resulting list can be used to visualize the NULL permutation distribution 
by passing it to \code{\link[CaDrA]{permutation_plot}}
See \code{\link[CaDrA]{CaDrA}} and \code{\link[CaDrA]{permutation_plot}} 
more details.
}
\usage{
data(perm.res)
}
\description{
A list of objects returned from \code{CaDrA()} when it runs
on simulated dataset (\code{ES = sim.ES}) with 
\code{input_score = ncol(sim.ES):1} and \code{method = "ks"}.
}
\examples{

# Load the pre-computed permutation results
data(perm.res)

# Plot the NULL permutation distribution against the observed best score
permutation_plot(perm.res)

}
\references{
Kartha VK, Kern JG, Sebastiani P, Zhang L,
Varelas X, Monti S (2017) CaDrA: A computational framework for performing 
candidate driver analyses using binary genomic features. 
(\href{https://www.frontiersin.org/articles/10.3389/fgene.2019.00121/full}{Frontiers in Genetics})
}
\keyword{datasets}
