% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm_data.R
\docType{data}
\name{perm.res}
\alias{perm.res}
\title{Permutation Results}
\format{
A list object returned by \code{topn.eval}
containing the ExpressionSet objects (and corresponding meta-feature scores)
for each search run when starting with each of the top 7 ranked features.
Returned ExpressionSet columns (i.e. samples) are ranked by the user-defined ranking variable.  
See \code{\link[CaDrA]{topn.eval}} for more information.
}
\usage{
data(perm.res)
}
\description{
The resulting list object returned by topn.eval() when run
on the \code{sim.ES} simulated dataset. All default parameters were used
when running topn.eval() except for max.size which was set to 10 
to account for the presence of 10 left-skewed (i.e. true positive or TP) 
features in the sim.ES dataset.
}
\examples{
data(perm.res)
}
\references{
Kartha VK, Kern JG, Sebastiani P, Zhang L,
Varelas X, Monti S (2017) CaDrA: A computational framework for performing 
candidate driver analyses using binary genomic features. 
(\href{https://www.biorxiv.org/content/early/2017/11/23/221846}{bioRxiv})
}
\keyword{datasets}
