% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_plot.R
\name{permutation_plot}
\alias{permutation_plot}
\title{Permutation Plot}
\usage{
permutation_plot(permutation_results)
}
\arguments{
\item{permutation_results}{a list of lists, where each list entry is one that is returned by the candidate search run for a given starting index. This is computed within and can be returned by the topn.eval() function.}
}
\value{
a density plot
}
\description{
Plots the Emperical Null distribution
}
\examples{
# Load pre-computed Top-N list generated for sim.ES dataset
data(topn.list)
topn_plot(topn.list)

}
