% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_plot.R
\name{permutation_plot}
\alias{permutation_plot}
\title{Plot the Empirical Null Distribution of Permutation Best Scores}
\usage{
permutation_plot(permutation_results)
}
\arguments{
\item{permutation_results}{a list of objects returned from the cadra_search()}
}
\value{
a density plot
}
\description{
Plot the Empirical Null Distribution of Permutation Best Scores
}
\examples{

# Load pre-computed permutation results
data(perm.res)

# Plot the permutation results
permutation_plot(perm.res)

}
