% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_plot.R
\name{permutation_plot}
\alias{permutation_plot}
\title{Permutation Best Scores Plot}
\usage{
permutation_plot(perm_res)
}
\arguments{
\item{perm_res}{a list of objects returned from CaDrA() function
using the simulated dataset (\code{sim.ES}) and random generated
\code{input_score = rnorm(n = ncol(sim.ES))} with set.seed(123).}
}
\value{
a density plot
}
\description{
Plot the Empirical Null Distribution of Permutation Best Scores Based on
a given number of permutations (\code{n_perm})
}
\examples{

# Load pre-computed permutation results
data(perm_res)

# Plot the permutation results
permutation_plot(perm_res)

}
