% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_plot.R
\name{plot_ESet}
\alias{plot_ESet}
\title{Enrichment Score (ES) plot}
\usage{
plot_ESet(df)
}
\arguments{
\item{df}{a data frame object with columns 'x' and 'y' containing the 
necessary data for an ES plot (returned by ks.genescore() 
when plot.dat is set to TRUE for a given dataset)}
}
\value{
A plot graphic of the Enrichment Score (ES) for a given distribution
}
\description{
Plot the ES running sum statistic, including the max score (KS) using x and 
y-axis data returned by ks.genescore() when plot.dat is set to TRUE 
for a given dataset
}
\examples{

# Load R library
library(Biobase)

# Load pre-computed Top-N list generated for sim.ES dataset
data(topn.list)

# With the results obtained from top-N evaluation,
# We can find the combination of features that gives 
# the best score in top N searches
topn_best_meta <- topn_best(topn_list=topn.list) 

# Extract the meta-feature set
ESet <-  topn_best_meta[["ESet"]]                    

# Make sure mat variable is a matrix
mat <- as.matrix(exprs(ESet))

# For cases when matrix only has one row
if(ncol(mat) == 1){
  mat <- matrix(t(mat), nrow=1, byrow=TRUE, 
  dimnames = list(rownames(ESet), rownames(mat))) 
}

# Add on the OR function of all the returned entries
or <- ifelse(colSums(mat)==0, 0, 1)
mat <- rbind(mat, or)

# Get x and y axis data for ES plot of 
# cumulative function of individual features (i.e. the OR function)
ES_dat <- ks_gene_score(
   n.x=length(or), y=which(or==1), plot_dat = TRUE, alternative = "less"
)

# plot for ES scores
ES_plot <- plot_ESet(df = ES_dat)

}
