% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revealer_genescore.R
\name{revealer_genescore}
\alias{revealer_genescore}
\title{Compute conditional mutual information scores of x and y given z}
\usage{
revealer_genescore(
  x,
  y,
  z = NULL,
  assoc_metric = "IC",
  target_match = "positive"
)
}
\arguments{
\item{x}{A continuous functional response of interest}

\item{y}{A binary feature for response of interest}

\item{z}{A binary feature which is often known as "causes" of activation}

\item{assoc_metric}{Association Metric: "IC" information coefficient (default) or "COR" correlation}

\item{target_match}{Direction of the match (negative or positive). Use "positive" to match the higher values of the target, "negative" to match the lower values. Default is positive.}
}
\description{
Compute conditional mutual information scores of x and y given z
}
