% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revealer_genescore_mat.R
\name{revealer_genescore_mat}
\alias{revealer_genescore_mat}
\title{\code{REVEALER} Scoring Method}
\usage{
revealer_genescore_mat(
  mat,
  target,
  target_match = "positive",
  seed_names = NULL,
  assoc_metric = c("IC", "COR"),
  verbose = FALSE
)
}
\arguments{
\item{mat}{a matrix of binary features (required)}

\item{target}{a vector of continuous values of a target profile (required). target must include labels or names that associated with the colnames of the feature matrix.}

\item{target_match}{a direction of target matching (\code{"negative"} or \code{"positive"}). Use \code{"positive"} to match the higher values of the target, \code{"negative"} to match the lower values. Default is \code{positive}.}

\item{seed_names}{one or more features(s) that associated with the activation of a given target profile}

\item{assoc_metric}{an association metric: \code{"IC"} information coefficient or \code{"COR"} correlation. Default is \code{IC}.}

\item{verbose}{a logical value indicates whether or not to print the diagnostic messages. Default is \code{FALSE}.}
}
\value{
a data frame with one column: \code{score}
}
\description{
Compute conditional mutual information of \code{x} and \code{y} given \code{z} for each row of a given binary feature matrix
}
\examples{

# Load R library
library(Biobase)

# Load pre-computed expression set
data(sim.ES)

# set seed
set.seed(123)

# Provide a vector of continuous scores for a target profile
target = rnorm(n = ncol(sim.ES))
names(target) <- colnames(sim.ES)

# Define additional parameters and start the function
revealer_genescore_result <- revealer_genescore_mat(
  mat = exprs(sim.ES), target = target, target_match = "positive", assoc_metric = "IC"
)
 
}
