% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revealer_genescore_mat.R
\name{revealer_genescore_mat}
\alias{revealer_genescore_mat}
\title{Row-wise matrix conditional mutual information I(x, y | z) from REVEALER}
\usage{
revealer_genescore_mat(
  mat,
  target,
  target_match = "positive",
  seed_names = NULL,
  seed_combination_op = "max",
  exclude_features = NULL,
  normalize_features = FALSE,
  count_thres_low = NULL,
  count_thres_high = NULL,
  assoc_metric = "IC",
  verbose = FALSE
)
}
\arguments{
\item{mat}{row matrix of binary features to compute row-wise scores for based on the REVEALER MUTUALLY EXCLUSIVE test}

\item{target}{a matrix of continuous functional response of interest}

\item{target_match}{Direction of the match (negative or positive). Use "positive" to match the higher values of the target, "negative" to match the lower values. Default is positive.}

\item{seed_names}{Starting seed, one or more binary features(s) representing known causes of activation or features associated with the target}

\item{seed_combination_op}{Operation to consolidate and summarize seeds to one vector of values. "max" is default}

\item{exclude_features}{Features to exclude in the search iterations}

\item{normalize_features}{Feature row normalization: FALSE (default) or "standardize" or "0.1.rescaling"}

\item{count_thres_low}{Filter out features with less than count.thres.low. NULL by default.}

\item{count_thres_high}{Filter out features with more than count.thres.high. NULL by default.}

\item{assoc_metric}{Assocication Metric: "IC" information coeff. (default) or "COR" correlation.}

\item{verbose}{a logical indicating whether or not to verbose diagnostic messages. Default is FALSE.}
}
\value{
A data frame
}
\description{
Compute conditional mutual information scores of x and y given z for each row of a given binary matrix
}
