% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{topn.eval}
\alias{topn.eval}
\title{Top 'N' evaluate}
\usage{
topn.eval(
  ranking = NULL,
  ESet,
  N = 1,
  do.plot = TRUE,
  best.score.only = TRUE,
  ...
)
}
\arguments{
\item{ranking}{integer vector specifying if and how samples should be re-ordered. Default is NULL. If NULL, we assume the ESet is already ordered}

\item{ESet}{an ordered expression set object with the same sample ordering and features as processed by the stepwise.search() function when performing a step-wise heuristic search}

\item{N}{an integer specifying the number of features to start the search over, starting from the top 'N' features in each case. Default is 1}

\item{do.plot}{a logical indicating whether you want to plot the resulting evaluation matrix. Default is TRUE}

\item{best.score.only}{a logical indicating whether or not to only return the best meta-feature score over the top 'N' evaluation. Default is TRUE}

\item{...}{additional parameters passed to the stepwise.search() function, which will be applied to each top 'N' run}
}
\value{
Default is a list of lists, where each list entry is one that is returned by the stepwise search run for a given starting index (See stepwise.search()). If best.score.only is set to TRUE, only the best score over the top N space is returned (useful for permutation-based testing)
1's and 0's represent whether a feature in any given row is present in a meta-feature along with a starting feature in the corresponding column.
}
\description{
Generates and evaluates stepwise search results for the top 'N' starting indices, checking for overlapping resulting features from each case. This function is mainly used to evaluate search results over the top 'N' best starting features for a given dataset.
}
