% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topn_best.R
\name{topn_best}
\alias{topn_best}
\title{Top 'N' Best Meta-Features}
\usage{
topn_best(topn_list)
}
\arguments{
\item{topn_list}{A nested list of objects that is returned from \code{candidate_search()}
by setting \code{FS = sim_FS}, \code{input_score = sim_Scores}, \code{top_N = 7},
\code{method = "ks_pval"}, \code{alternative = "less"}, \code{search_method = "both"}, 
\code{max_size = 10}, and \code{best_score_only = FALSE}. The nested list contains 
the meta-feature sets as well as the observed input scores and its corresponding 
best score for each top 'N' search.}
}
\value{
A list of objects containing the best meta-feature, its corresponding best score, 
and observed input scores
}
\description{
Take the resulting list of meta-features returned from \code{candidate_search()}
and fetch the meta-feature with the best score
}
\examples{

# Load pre-computed Top-N list generated for sim_FS dataset
data(topn_list)

# Get the best meta-features list
topn_best_meta <- topn_best(topn_list = topn_list)

}
