% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topn_list.R
\docType{data}
\name{topn_list}
\alias{topn_list}
\title{Top-N Results for Simulated Data (\code{sim_ES})}
\format{
A list of objects returned from \code{candidate_search()} function
containing a set of SummarizedExperiment objects, input_score, corresponding best score,
method and metric used for each top_N feature searches.
See \code{\link[CaDrA]{candidate_search}} for more information.
}
\usage{
data(topn_list)
}
\description{
A list of objects returned from \code{candidate_search()} using simulated dataset
\code{ES = sim_ES}, \code{input_score = ncol(sim_ES):1}, \code{top_N = 7},
\code{method = "ks"}, \code{alternative = "less"}, \code{metric = "pval"},
\code{search_method = "both"}, \code{max_size = 10},
and \code{best_score_only = FALSE} as inputs to the function.
\code{Note:} \code{max_size} is set to 10 as we would like to account
for the presence of 10 left-skewed
(i.e. true positive or TP) features in the \code{sim_ES} dataset.
Over top_N = 7 feature searches, a list of SummarizedExperiment objects, input_score,
along with its corresponding best score, method and metric are returned from each search.
}
\examples{

# Load pre-computed Top-N list generated for sim_ES and sim_Scores dataset
data(topn_list)

# For N=1, The result is initiated with the top-scoring starting feature
topn_list[[1]]

# Fetch the SummarizedExperiment object for the first-best feature
topn_list[[1]]$eset

# Fetch the SummarizedExperiment object for the second-best feature
topn_list[[2]]$eset

# Retrieve the result of best scores over top_N = 7 runs
topn_best_meta <- topn_best(topn_list = topn_list)

# Visualize the best meta-feature using meta_plot() function
meta_plot(topn_best_list = topn_best_meta)

# Visualize overlap of features across the top_N = 7 using topn_plot() function
topn_plot(topn_list = topn_list)

}
\references{
Kartha VK, Kern JG, Sebastiani P, Zhang L,
Varelas X, Monti S (2019) CaDrA: A computational framework for performing
candidate driver analyses using binary genomic features.
(\href{https://www.frontiersin.org/articles/10.3389/fgene.2019.00121/full}{Frontiers in Genetics})
}
\keyword{datasets}
