% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topn_plot.R
\name{topn_plot}
\alias{topn_plot}
\title{Top 'N' Plot}
\usage{
topn_plot(topn_list, verbose = FALSE)
}
\arguments{
\item{topn_list}{a list of lists where each list entry is returned from the \code{candidate_search()} for a given starting index. This is computed within and can be returned by the \code{topn_eval()} function.}

\item{verbose}{a logical value indicates whether or not to print the diagnostic messages. Default is \code{FALSE}.}
}
\value{
a heatmap of the top N evaluation for a given top N search evaluation
}
\description{
Plots a heatmap representation of overlapping features given a list of top N features obtained from \code{candidate_search()} results
}
\examples{

# Load pre-computed Top-N list generated for sim.ES dataset
data(topn.list)

# Get top N plot
topn_plot(topn.list)

}
