% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcox_genescore.R
\name{wilcox_genescore}
\alias{wilcox_genescore}
\title{Compute rank sum scores for a given binary feature}
\usage{
wilcox_genescore(
  x,
  y,
  mu = 0,
  alternative = "less",
  paired = FALSE,
  exact = FALSE,
  correct = TRUE
)
}
\arguments{
\item{x}{an integer ranked values for group 1}

\item{y}{an integer ranked values for group 2}

\item{mu}{a number uses as an optional parameter to form a null hypothesis. Default is \code{0}.}

\item{alternative}{alternative hypothesis for p-value calculation (\code{"two.sided"} or \code{"greater"} or \code{"less"}). Default is \code{less} for left-skewed significance testing.}

\item{paired}{whether to perform paired test. Default is \code{FALSE}.}

\item{exact}{whether to compute exact p-value. Default is \code{FALSE}.}

\item{correct}{whether to consider continuity correction for p-value. Default is \code{TRUE}.}
}
\value{
a data frame with two columns: \code{score} and \code{p_value}
}
\description{
Compute rank sum scores for a given binary feature
}
