% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcox_genescore.R
\name{wilcox_genescore}
\alias{wilcox_genescore}
\title{Compute rank sum scores for each row of a given binary matrix}
\usage{
wilcox_genescore(
  x,
  y,
  mu = 0,
  alternative = "less",
  paired = FALSE,
  exact = FALSE,
  correct = TRUE
)
}
\arguments{
\item{x}{ranks for group 1}

\item{y}{ranks for group 2}

\item{mu}{a number specifying an optional parameter used to form null hypothesis}

\item{alternative}{alternative hypothesis for p-value calculation}

\item{paired}{paired test}

\item{exact}{compute exact p-value}

\item{correct}{for continuity correction (p-value)}
}
\value{
A data frame
}
\description{
Compute rank sum scores for each row of a given binary matrix
}
