% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcox_genescore.R
\name{wilcox_genescore}
\alias{wilcox_genescore}
\title{Compute rank sum scores for a given binary feature}
\usage{
wilcox_genescore(
  x,
  y,
  mu = 0,
  alternative = c("less", "greater", "two.sided"),
  paired = FALSE,
  exact = FALSE,
  correct = TRUE
)
}
\arguments{
\item{x}{an integer ranked values for group 1}

\item{y}{an integer ranked values for group 2}

\item{mu}{a number uses as an optional parameter to form a null hypothesis. 
Default is \code{0}.}

\item{alternative}{alternative hypothesis for p-value calculation 
(\code{"two.sided"} or \code{"greater"} or \code{"less"}). 
Default is \code{less} for left-skewed significance testing.}

\item{paired}{whether to perform paired test. Default is \code{FALSE}.}

\item{exact}{whether to compute exact p-value. Default is \code{FALSE}.}

\item{correct}{whether to consider continuity correction for p-value. 
Default is \code{TRUE}.}
}
\value{
a data frame with two columns: \code{score} and \code{p_value}
}
\description{
Compute rank sum scores for a given binary feature
}
\examples{

# Load R library
library(Biobase)
# Load pre-computed expression set
data(sim.ES)

# set seed
set.seed(123)

# Provide a vector of continuous scores for a target profile 
# The scores must have labels or names 
# that match the colnames of expression matrix
input_score = rnorm(n = ncol(sim.ES))
names(input_score) <- colnames(sim.ES)

# Sort input_score from highest to lowest values
input_score <- sort(input_score, decreasing=TRUE)

# Extract the expression matrix
mat <- exprs(sim.ES)

# Re-order the samples by input_score sorted from highest to lowest values
mat <- mat[,names(input_score)]

# Define ranks of input_score
ranks <- seq_along(input_score)

# Define alternative
alternative <- "less"

# Compute the wilcox rank sum statitic and p-value per row in the matrix
wilcox <- apply(X=mat, MARGIN=1, function(x, r=ranks){
 wilcox_genescore(x=r[which(x==1)], y=r[which(x==0)], 
 alternative=alternative) 
})

}
