% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcox_genescore_mat.R
\name{wilcox_genescore_mat}
\alias{wilcox_genescore_mat}
\title{Row-wise matrix Wilcoxon rank sum scoring}
\usage{
wilcox_genescore_mat(
  mat,
  ranks = NULL,
  alternative = c("two.sided", "greater", "less"),
  verbose = FALSE
)
}
\arguments{
\item{mat}{matrix of binary features to compute row-wise scores for based on the Wilcoxon rank sum test}

\item{ranks}{a vector of ranks to use when performing the Wilcoxon test. Default is NULL. If NULL, then samples are assumed to be ordered by increasing ranking. Value passed to wilcox_genescore() function}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided","less" or "greater". Value passed to wilcox_genescore() function}

\item{verbose}{a logical indicating whether or not to verbose diagnostic messages. Default is FALSE}
}
\value{
A data frame with two columns: \code{score} and \code{p_value}
}
\description{
Row-wise matrix Wilcoxon rank sum scoring
}
\examples{
# Load R library
library(Biobase)
# Load pre-computed expression set
data(sim.ES)

# Compute the score
wilcox_genescore_mat_result <- wilcox_genescore_mat(
  mat = exprs(sim.ES), 
  ranks = NULL,
  alternative = "less"
)

}
